+++ artifact-evaluation/bundle-ae.sh
@@ -2,6 +2,11 @@
 # Convenience script to package an artifact for TACO, designed to be run with 
 # podman
 
+# Strict mode with error reporting
+set -euo pipefail
+trap 's=$?; echo "$0: Error on line "$LINENO": $BASH_COMMAND"; exit $s' ERR
+IFS=$'\n\t'
+
 # Remove all unnecessary stuff first
 # Remove build output
 cargo clean
@@ -11,12 +16,14 @@ rm -rf docs/_build
 mkdir artifact
 
 # Artifact README
-cp resources/artifact-evaluation.md artifact/README.md
+cp artifact-evaluation/artifact-evaluation.md artifact/README.md
 # LICENSE
 cp ./LICENSE artifact/LICENSE
 
 # Paper pdf
-#cp resources/paper_5548.pdf artifact/paper_5548.pdf
+#cp artifact-evaluation/paper.pdf artifact/paper.pdf
+# Table 2 ods
+cp artifact-evaluation/table2.ods artifact/table2.ods
 
 # Create source directory
 mkdir artifact/src